+++ TChapter/Util/ConvertUtil.cs
@@ -51,7 +51,7 @@ namespace TChapter.Util
         {
             var self = data[index];
             return self.Chapters.Where(c => c.Time != TimeSpan.MinValue)
-                .Select(c => ((long) Math.Round(c.Time.TotalSeconds * self.FramesPerSecond)).ToString()).ToArray();
+                .Select(c => ((long) Math.Round((decimal) c.Time.TotalSeconds * self.FramesPerSecond)).ToString()).ToArray();
         }
 
         public static string[] ToTIMECODES(this MultiChapterData data, int index = 0)
